backend/requirements.txt
# FastAPI and dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
asyncpg==0.30.0
psycopg2-binary==2.9.9
alembic==1.13.1

# Redis
redis==5.0.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
PyJWT==2.8.0

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.2.0

# File handling
aiofiles==23.2.1
pillow==11.0.0

# WebSocket
websockets==12.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Development
python-dotenv==1.0.0